Fix analytics
